Mayor Update! V1.63b - Added AMBE-addon modules - Added EEProm / Config-Storage support - split main file 'rptr_main.c' into 3 modules:   'rptr_main.c': init + main-loop   'config.c'  : load/save config blocks from/to EEProm, configure functions   'stdmodem.c' : "behavior" functions, communication with PC
git-svn-id: svn://svn.code.sf.net/p/dv-rptr-fw/code/trunk@79 208d634f-0f76-4b8c-b329-1362b4586a7b

+++ src/opendv/dv_dstar.c
@@ -40,7 +40,7 @@
 
 
 
-void dstar_buildheader(unsigned long *wrbuffer, const tds_header *header) {
+void dstar_buildheader(unsigned int *wrbuffer, const tds_header *header) {
   int hdr_bitcnt, hdr_bitpos, hdr_bytcnt; //, hdr_interbit;
   char *hdr_rptr = (char *)header;	// Pointer to Original Header
   char *hdr_wbuf = (char *)wrbuffer;	// Byte-Pointer to storage buffer
